Make sure recorded delta associated with MouseWheel are +/-120
The delta associated with Qt Mouse Wheel is sometimes greater than 120
(for example when the wheel is turned very fast). This is expected.
From our observation, it seems additional event are also emitted/recorded.

Further investigation are needed to exactly identified the issue.
The CTK issue #132 will allow us to track that problem.